+++ pycognaize/__init__.py
@@ -2,12 +2,13 @@
     all the data and functionality provided by the pycognaize application.
 """
 
-__version__ = "1.1.2"
+__version__ = "1.2.5"
 
-__all__ = ['Model', 'Snapshot']
+__all__ = ['Login', 'Model', 'Snapshot']
 
 import os
 from .model import Model
+from .login import Login
 from .document import Snapshot
 
 name = os.path.basename(os.path.dirname(__file__))