+++ pycognaize/common/enums.py
@@ -29,6 +29,7 @@ class EnvConfigEnum(enum.Enum):
     # IQ_DATASNAP_URL = "IQ_DATASNAP_URL"
     SNAPSHOT_ID = 'SNAPSHOT_ID'
     SNAPSHOT_PATH = "SNAPSHOT_PATH"  # Images, OCR and snap.json
+    HOST = "COGNAIZE_HOST"
 
 
 class IqCollectionEnum(enum.Enum):
@@ -60,6 +61,7 @@ class IqDocumentKeysEnum(enum.Enum):
     tags = 'tags'
     pages = 'pages'
     name = 'name'
+    section = 'section'
     field_type = 'fieldType'
     field_id = '_id'
     src_field_id = "srcFieldId"