+++ pycognaize/document/field/section_field.py
@@ -1,46 +1,31 @@
-"""
-| TextField objects represent a single labeled pycognaize field.
-| The latter can have no tags, a single tag or multiple tags.
-
->>> from pycognaize.document.field.text_field import TextField
->>> text_field = TextField(name='company name', value='gooloogooloo')
->>> text_field.name
-'company name'
->>> text_field.value
-'gooloogooloo'
->>> text_field.tags
-[]
-"""
 import logging
 from typing import List, Optional, Dict, Type
 
 
 from pycognaize.common.enums import (
     IqDocumentKeysEnum,
-    IqTagKeyEnum,
     ID,
     IqFieldKeyEnum,
     IqDataTypesEnum
 )
 from pycognaize.document.page import Page
 from pycognaize.document.field import Field
-from pycognaize.document.tag import ExtractionTag
+from pycognaize.document.tag.section_tag import SectionTag
 
 
-class TextField(Field):
-    """Base class for all pycognaize text fields"""
-    tag_class: Type[ExtractionTag] = ExtractionTag
+class SectionField(Field):
+    """Base class for all pycognaize Section fields"""
+    tag_class: Type[SectionTag] = SectionTag
 
     def __init__(self,
                  name: str,
-                 value: str = '',
-                 tags: Optional[List[ExtractionTag]] = None,
+                 tags: Optional[List[SectionTag]] = None,
                  field_id: Optional[str] = None,
                  group_key: str = None,
                  confidence: Optional[float] = -1.0,
                  group_name: str = None
                  ):
-        """ Create a TextField object
+        """ Create a SectionField object
 
         :param name: Name of the field
         :param value: Value of the field
@@ -48,31 +33,29 @@ class TextField(Field):
         :param tags: List of tag objects
         :param field_id: The id of the field
         """
-        super().__init__(name=name, tags=tags, value=value,
+        super().__init__(name=name, tags=tags,
                          group_key=group_key, confidence=confidence,
                          group_name=group_name)
         self._field_id = field_id
-        self._value = '; '.join([i.raw_value
-                                 for i in self.tags]) if self.tags else value
-
-    @property
-    def value(self):
-        return self._value
 
     @classmethod
     def construct_from_raw(
-            cls, raw: dict, pages: Dict[int, Page]) -> 'TextField':
-        """Create TextField object from dictionary"""
-        tag_dicts: List[dict] = raw[IqDocumentKeysEnum.tags.value]
+            cls, raw: dict, pages: Dict[int, Page]) -> 'SectionField':
+        """Create SectionField object from dictionary"""
+        section_dict: List[dict] = raw[IqDocumentKeysEnum.tags.value]
         tags = []
-        for i in tag_dicts:
-            try:
-                tags.append(cls.tag_class.construct_from_raw(
-                    raw=i, page=pages[i['page']]))
-            except Exception as e:
-                logging.debug(f"Failed creating tag for field {raw[ID]}: {e}")
+        for i in section_dict:
+            for tag_type, tag_data in i[IqDocumentKeysEnum.
+                                        section.value].items():
+                try:
+                    tags.append(cls.tag_class.construct_from_raw(
+                                raw=tag_data,
+                                pages=pages,
+                                tag_type=tag_type))
+                except Exception as e:
+                    logging.debug(f"Failed creating tag"
+                                  f" for field {raw[ID]}: {e}")
         return cls(name=raw[IqDocumentKeysEnum.name.value],
-                   value=raw[IqTagKeyEnum.value.value],
                    tags=tags,
                    field_id=str(raw[ID]),
                    group_key=raw.get(IqFieldKeyEnum.group_key.value, ''),
@@ -88,8 +71,7 @@ class TextField(Field):
         return field_dict
 
     def __repr__(self):
-        return (f"<{self.__class__.__name__}: {self.name}:"
-                f" {'|'.join([i.raw_value for i in self.tags])}>")
+        return f"<{self.__class__.__name__}: {self.name}>"
 
     def __str__(self):
-        return f"{'|'.join([i.raw_value for i in self.tags])}"
+        return f"<{self.__class__.__name__}: {self.name}>"