+++ CHANGELOG.md
@@ -1,6 +1,37 @@
 # Changelog
 
-## [1.0]
+## [1.2]
+### [1.2.5] - 2022-11-28
+- Read page image height/width from document.json
+- Field raw value bug fix
+
+### [1.2.4] - 2022-11-28
+- Field object raw_value bug fix
+
+### [1.2.3] - 2022-11-16
+- Field object raw_value bug fix
+
+### [1.2.3] - 2022-11-16
+- Field object raw_value contains fields value
+
+### [1.2.2] - 2022-11-11
+- Add HTTP request timeout for genie model run
+
+### [1.2.1] - 2022-11-08
+- Add functionality for grouping fields by key
+
+
+### [1.2.0] - 2022-10-11
+- Modify assign indices functionality to correctly index tables located side by side
+
+## [1.1]
+### [1.1.4] - 2022-10-06
+- Add snapshot download to specified directory
+  functionality
+- Add Page Section tag and field functionalities
+
+### [1.1.3] - 2022-10-05
+- Add login functionality to pycognaize
 
 ### [1.1.2] - 2022-09-12
 - group_by_field returns list of fields with group_key
@@ -15,6 +46,7 @@
 - Added multiprocessing download of images and 
   ocr data
 
+## [1.0]
 ### [1.0.3] - 2022-08-23
 - Change get_tied_fields, get_tied_tags, 
   get_first_tied_field, get_first_tied_tag methods to return also python names